useradv listo

+++ server/controllers/user.controles.js
@@ -1,12 +1,17 @@
 import { json } from "express";
 import { pool } from "../db.js";
+import bcrypt from "bcrypt";
 
 async function getUsers(req, res) {
   try {
-    const [result] = await pool.query(
-      "SELECT id, avatar_img, nombres, apellidos, rut, username FROM user"
-    );
-    res.json(result);
+    if(req.session.user.es_admin){
+      const [result] = await pool.query(
+        "SELECT id, avatar_img, nombres, apellidos, rut, username FROM user"
+      );
+      res.json(result);
+    }else{
+      return res.status(404).json({ message: "No tienes permisos" });
+    }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
   }
@@ -14,14 +19,18 @@ async function getUsers(req, res) {
 
 async function getoneUser(req, res) {
   try {
-    const [result] = await pool.query(
-      "SELECT * FROM user WHERE id = ?",
-      req.params.id
-    );
-    if (result.length === 0) {
-      return res.status(404).json({ Message: "User not found" });
-    } else {
-      res.json(result[0]);
+    if(req.session.user.es_admin){
+      const [result] = await pool.query(
+        "SELECT * FROM user WHERE id = ?",
+        req.params.id
+      );
+      if (result.length === 0) {
+        return res.status(403).json({ Message: "User not found" });
+      } else {
+        res.json(result[0]);
+      }
+    }else{
+      return res.status(404).json({ message: "No tienes permisos" });
     }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
@@ -30,21 +39,8 @@ async function getoneUser(req, res) {
 
 async function createUsers(req, res) {
   try {
-    const {
-      nombres,
-      apellidos,
-      rut,
-      username,
-      email,
-      password,
-      es_admin,
-      direccion,
-      telefono,
-      avatar_img,
-    } = req.body;
-    const [result] = await pool.query(
-      "INSERT INTO user (nombres, apellidos, rut, username, email, password, es_admin, direccion, telefono, avatar_img) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-      [
+    if(req.session.user.es_admin){
+      const {
         nombres,
         apellidos,
         rut,
@@ -55,12 +51,33 @@ async function createUsers(req, res) {
         direccion,
         telefono,
         avatar_img,
-      ]
-    );
-    res.json({
-      id: result.insertId,
-      message: `Usuario creado exitosamente`,
-    });
+      } = req.body;
+
+      const salt = await bcrypt.genSalt(10);
+      const hash = bcrypt.hashSync(password, salt);
+
+      const [result] = await pool.query(
+        "INSERT INTO user (nombres, apellidos, rut, username, email, password, es_admin, direccion, telefono, avatar_img) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
+        [
+          nombres,
+          apellidos,
+          rut,
+          username,
+          email,
+          hash,
+          es_admin,
+          direccion,
+          telefono,
+          avatar_img,
+        ]
+      );
+      res.json({
+        id: result.insertId,
+        message: `Usuario creado exitosamente`,
+      });
+  }else{
+    return res.status(404).json({ message: "No tienes permisos" });
+  }
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
@@ -68,13 +85,17 @@ async function createUsers(req, res) {
 
 async function uptadateUsers(req, res) {
   try {
-    const [result] = await pool.query("UPDATE user SET ? WHERE id = ?", [
-      req.body,
-      req.params.id,
-    ]);
-    if (result.affectedRows === 0)
-      return res.status(404).json({ message: "No se actualizo ningún dato" });
-    res.json(result);
+    if (req.session.user.es_admin || req.params.id == req.session.user.user_id) {
+      const [result] = await pool.query("UPDATE user SET ? WHERE id = ?", [
+        req.body,
+        req.params.id,
+      ]);
+      if (result.affectedRows === 0)
+        return res.status(403).json({ message: "No se actualizo ningún dato" });
+      res.json(result);
+    } else {
+      return res.status(404).json({ message: "No tienes permisos" });
+    }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
   }
@@ -82,34 +103,72 @@ async function uptadateUsers(req, res) {
 
 async function deleteUsers(req, res) {
   try {
-    const [result] = await pool.query(
-      "DELETE FROM user WHERE id = ?",
-      req.params.id
-    );
-    if (result.affectedRows === 0)
-      return res.status(404).json({ message: "Usuario no encontrado" });
-    return res.status(204).json({ message: "Usuario eliminado exitosamente" });
+    if (req.session.user.es_admin) {
+      if (req.params.id == req.session.user.user_id) {
+        return res
+          .status(403)
+          .json({ message: "No puedes eliminarte a ti mismo" });
+      }
+      const [result] = await pool.query(
+        "DELETE FROM user WHERE id = ?",
+        req.params.id
+      );
+      if (result.affectedRows === 0)
+        return res.status(403).json({ message: "Usuario no encontrado" });
+      return res
+        .status(204)
+        .json({ message: "Usuario eliminado exitosamente" });
+    } else {
+      return res.status(404).json({ message: "No tienes permisos" });
+    }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
   }
 }
 
-async function login(req, res){
-  try{
+async function login(req, res) {
+  try {
     const { username, password } = req.body;
 
-    const [result] = await pool.query(
-      "SELECT * FROM user WHERE username = ? and password = ?",
-      [ username, password]
-    )
-    if (result.length === 0) {
-      return res.status(404).json({ Message: "Credenciales erroneas" });
+    if (req.session.authenticated) {
+      res.json(req.session);
     } else {
-      return res.status(200).json({ message: "Inicio de sesión exitoso"});
+      const [result] = await pool.query(
+        "SELECT * FROM user WHERE username = ?",
+        [username]
+      );
+
+      const auth = bcrypt.compareSync(password, result[0].password);
+
+      if (auth) {
+        req.session.authenticated = true;
+        const user_id = result[0].id;
+        const es_admin = result[0].es_admin;
+        req.session.user = {
+          user_id,
+          username,
+          es_admin,
+        };
+        return res.json(req.session);
+      } else {
+        return res.status(403).json({ message: "Usuario no encontrado" });
+      }
     }
-  }
-  catch(error){
+  } catch (error) {
     return res.status(500).json({ message: console.error.message });
+  }
+}
+
+async function logout(req, res) {
+  if (req.session.authenticated) {
+    req.session.destroy((err) => {
+      if (err) {
+        console.error(err);
+      }
+      return res.status(200).json({ message: "Sesión cerrada" }); // Redirige al inicio de sesión después de cerrar la sesión
+    });
+  } else {
+    return res.status(403).json({ message: "Usuario no iniciado" }); // Si el usuario no está autenticado
   }
 }
 
@@ -120,4 +179,5 @@ export default {
   uptadateUsers,
   deleteUsers,
   login,
+  logout,
 };
